added resizing (kind of)

+++ imgui/imgui_widgets.cpp
@@ -3567,7 +3567,7 @@ bool ImGui::InputFloat4(const char* label, float v[4], const char* format, ImGui
     return InputScalarN(label, ImGuiDataType_Float, v, 4, NULL, NULL, format, flags);
 }
 
-bool ImGui::InputInt(const char* label, long long int* v, long long int step, int step_fast, ImGuiInputTextFlags flags)
+bool ImGui::InputInt(const char* label, long long int* v, long long int step, long long int step_fast, ImGuiInputTextFlags flags)
 {
     // Hexadecimal input provided as a convenience but the flag name is awkward. Typically you'd use InputText() to parse your own data, if you want to handle prefixes.
     const char* format = (flags & ImGuiInputTextFlags_CharsHexadecimal) ? "%08X" : "%lld";
